numeric: fix Richardson by running on masses and values separately
